requirements.txt
# Core ML and Data Processing
pandas==2.1.4
numpy==1.26.2
scikit-learn==1.3.2
joblib==1.3.2

# MLflow and Experiment Tracking
mlflow==2.9.2
dagshub==0.3.19

# Data Version Control
dvc==3.38.1
dvc-s3==3.1.0

# API and Web Framework
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.2
requests==2.31.0

# Data Quality and Profiling
pandas-profiling==3.6.6
great-expectations==0.18.8

# Monitoring
prometheus-client==0.19.0

# Airflow
apache-airflow==2.7.3
apache-airflow-providers-docker==3.7.0

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1

# Testing
pytest==7.4.3
pytest-cov==4.1.0
black==23.11.0
flake8==6.1.0

